First attempt at abstracting SSH handling
Creates functions for working with SSH connections and channels, i.e.
openening, closing and executing commands.

Adds wrapper structs around the Erlang :ssh connection references and
channel ids that allow storing the initial config/options as well as
other references.

SSH channel messages are forwarded to a handler function, together with
the `%Channel` wrapper object, the message and a state, which can be
transformed in each handler invocation. See `SSHKit.SSH.run/4` and
`SSHKit.SSH.capture/3` for an example.

Adds a quick sanity check running against the Docker container.
You'll need to run the container in the background yourself for now.