[#72] Allow SSH.connect to receive a lambda where the connection is only open within the lambda
This extends SSHKit.SSH with connect/3 which accepts a function that is
called with the connection. The connection is automatically closed after
the function returned, no matter if it errored or not.

Also: this introduces a new option - "ssh_modules" for
SSHKit.SSH.connect that allows to use own implementations for the
underlying Erlang modules. This can be used for unit testing SSHKit.

close #72