[#76] Add functional tests for SSH.connect with lambda functions (#77)
* [#76] Add functional tests for SSH.connect with lambda functions

This adds another small functional test for SSH.connect. It does the
same as the existing one but utilizes a lambda function for this.

* Fix flaky functional tests for SSHKit.SSH

The tests need to use the `silently_accept_hosts` option for Erlang's
ssh module, otherwise user interaction would be necessary to add new
hosts to the known_hosts file.

close #76

* Refactor functional tests for SSHKit.SCP

Update tests to use the newly introduced lambda-syntax (is this the
Elixir-way of saying this?).