Avoid changing directories in functional tests
Use of `File.cd` is a problem if we wish to run functional tests in
parallel as it mutates shared process state. The current working
directory for all test processes is updated which means local
file-system operations are executed in a different context.

E.g. if we `File.cd` in one test and another test, running in parallel,
tries to calculate the `shasum` of a file in its current directory, then
these tests interfere.